% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api_token_.R
\name{check_api_token_}
\alias{check_api_token_}
\title{Checking API Token}
\usage{
check_api_token_(api_token)
}
\arguments{
\item{api_token}{To validate your requests, you'll need your api_token - this means that our system will need to know who you are and be able to connect all actions you do with your chosen Pipedrive account. Have in mind that a user has a different api_token for each company. Please access the following link for more information: https://pipedrive.readme.io/docs/how-to-find-the-api-token?utm_source=api_reference}
}
\value{
string
}
\description{
Internal function to check API Token
}
\examples{
\dontrun{
#Example
check_api_token_("sjdfsanm23423")
}
}
