% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_list_.R
\name{clear_list_}
\alias{clear_list_}
\title{Clear List}
\usage{
clear_list_(x)
}
\arguments{
\item{x}{List that you wish remove fields with NULL}
}
\value{
a List wihtout fileds wihout
}
\description{
Remove fields NULL in a list
}
\examples{
\dontrun{
#Example
clear_list_(list(field1 = 123, field2 = NULL, field3 = "name"))
}
}
