% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealfields.get_all.R
\name{dealfields.get_all}
\alias{dealfields.get_all}
\title{Get all deal fields}
\usage{
dealfields.get_all(start = NULL, limit = NULL, api_token = NULL,
  company_domain = "api")
}
\arguments{
\item{start}{Pagination start}

\item{limit}{Items shown per page}

\item{api_token}{To validate your requests, you'll need your api_token - this means that our system will need to know who you are and be able to connect all actions you do with your chosen Pipedrive account. Have in mind that a user has a different api_token for each company. Please access the following link for more information: <https://pipedrive.readme.io/docs/how-to-find-the-api-token?utm_source=api_reference>}

\item{company_domain}{How to get the company domain: <https://pipedrive.readme.io/docs/how-to-get-the-company-domain>}
}
\value{
return an object List()
}
\description{
Function to Get all deal fields from pipedrive.
}
\examples{
\donttest{
dealfields.get_all(api_token='token',company_domain='exp')
}
}
