% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.get.download.R
\name{files.get.download}
\alias{files.get.download}
\title{Download one file}
\usage{
files.get.download(id, save, api_token = NULL, company_domain = "api")
}
\arguments{
\item{id}{ID of the file to download.}

\item{save}{File name that you wish save, the default path is Work Directory, you can put full path (path + file name) to save where you want}

\item{api_token}{To validate your requests, you'll need your api_token - this means that our system will need to know who you are and be able to connect all actions you do with your chosen Pipedrive account. Have in mind that a user has a different api_token for each company. Please access the following link for more information: <https://pipedrive.readme.io/docs/how-to-find-the-api-token?utm_source=api_reference>}

\item{company_domain}{How to get the company domain: <https://pipedrive.readme.io/docs/how-to-get-the-company-domain>}
}
\value{
return an object List()
}
\description{
Function to Download one file from pipedrive.
}
\examples{
\donttest{
files.get.download(id='e.g.',save='e.g.',api_token='token',company_domain='exp')
}
}
