% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_.R
\name{get_all_}
\alias{get_all_}
\title{Execute GET from pipedrive's URL}
\usage{
get_all_(url)
}
\arguments{
\item{url}{Pipedrive URL for get informations from bases}
}
\value{
a data frame with result
}
\description{
Execute GET from pipedrive's URL, download all pages from URL and return a data frame with result.
}
\examples{
\dontrun{
#Example
get_all_("https://api.pipedrive.com/v1/activities?start=0&api_token=exemaple_token")
}
}
