% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalmessages.get.R
\name{globalmessages.get}
\alias{globalmessages.get}
\title{Get global messages}
\usage{
globalmessages.get(limit = NULL, api_token = NULL,
  company_domain = "api")
}
\arguments{
\item{limit}{Number of messages to get from 1 to 100. 1 message is returned by default.}

\item{api_token}{To validate your requests, you'll need your api_token - this means that our system will need to know who you are and be able to connect all actions you do with your chosen Pipedrive account. Have in mind that a user has a different api_token for each company. Please access the following link for more information: <https://pipedrive.readme.io/docs/how-to-find-the-api-token?utm_source=api_reference>}

\item{company_domain}{How to get the company domain: <https://pipedrive.readme.io/docs/how-to-get-the-company-domain>}
}
\value{
return an object List()
}
\description{
Function to Get global messages from pipedrive.
}
\examples{
\donttest{
globalmessages.get(api_token='token',company_domain='exp')
}
}
