% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.get.permittedusers.R
\name{organizations.get.permittedusers}
\alias{organizations.get.permittedusers}
\title{List permitted users}
\usage{
organizations.get.permittedusers(id, access_level = NULL,
  api_token = NULL, company_domain = "api")
}
\arguments{
\item{id}{ID of the organization}

\item{access_level}{If set, filter results by allowed access level. 1 = Read, 2 = Write, 3 = Read+Write This field has the following domains: (1; 2; 3)}

\item{api_token}{To validate your requests, you'll need your api_token - this means that our system will need to know who you are and be able to connect all actions you do with your chosen Pipedrive account. Have in mind that a user has a different api_token for each company. Please access the following link for more information: <https://pipedrive.readme.io/docs/how-to-find-the-api-token?utm_source=api_reference>}

\item{company_domain}{How to get the company domain: <https://pipedrive.readme.io/docs/how-to-get-the-company-domain>}
}
\value{
return an object List()
}
\description{
Function to List permitted users from pipedrive.
}
\examples{
\donttest{
organizations.get.permittedusers(id='e.g.',api_token='token',company_domain='exp')
}
}
