% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_url_parameters_.R
\name{prepare_url_parameters_}
\alias{prepare_url_parameters_}
\title{Transform list to string}
\usage{
prepare_url_parameters_(x)
}
\arguments{
\item{x}{list with parameters that you wish put in url format request}
}
\value{
a formatted string
}
\description{
Transform list in string, for use in requests from pipedrive's API
}
\examples{
\dontrun{
#Example
prepare_url_parameters_(list(field1 = 123, field2 = NULL, field3 = "name"))
}
}
