% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.get_all.R
\name{stages.get_all}
\alias{stages.get_all}
\title{Get all stages}
\usage{
stages.get_all(pipeline_id = NULL, api_token = NULL,
  company_domain = "api")
}
\arguments{
\item{pipeline_id}{ID of the pipeline to fetch stages for. If omitted, stages for all pipelines will be fetched.}

\item{api_token}{To validate your requests, you'll need your api_token - this means that our system will need to know who you are and be able to connect all actions you do with your chosen Pipedrive account. Have in mind that a user has a different api_token for each company. Please access the following link for more information: <https://pipedrive.readme.io/docs/how-to-find-the-api-token?utm_source=api_reference>}

\item{company_domain}{How to get the company domain: <https://pipedrive.readme.io/docs/how-to-get-the-company-domain>}
}
\value{
return an object List()
}
\description{
Function to Get all stages from pipedrive.
}
\examples{
\donttest{
stages.get_all(api_token='token',company_domain='exp')
}
}
